[7] modified cards
