+++ frontend/src/components/Cards/card.module.css
@@ -1,0 +1,198 @@
+.card__block {
+  width: 820px;
+  margin: 0 auto;
+}
+
+/*.card {
+    position: relative;
+    display: block;
+    height: 480px;
+    cursor: pointer;
+  }
+  
+  .front,
+  .back {
+    position: absolute;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+    border: 3px solid #000;
+    padding: 20px;
+    overflow: hidden;
+    background-color: #fdfdfd;
+    transition: transform .8s linear;
+    backface-visibility: hidden;
+  }
+  
+  .front {
+    transform: perspective(1200px) rotateY(0deg);
+  }
+  
+  .back {
+    transform: perspective(1200px) rotateY(180deg);
+  }
+  
+  .card:hover>.front {
+    transform: perspective(1200px) rotateY(-180deg);
+  }
+  
+  .card:hover>.back {
+    transform: perspective(1200px) rotateY(0deg);
+  }
+  
+  .front__title {
+    margin: 0;
+    color: #000;
+    font-weight: 400;
+    font-size: 64px;
+    text-align: center;
+  }
+  
+  .back__description {
+    margin: 0;
+    color: #000;
+    font-weight: 400;
+    font-size: 32px;
+    text-align: center;
+  }*/
+
+label {
+  position: relative;
+  display: block;
+  width: 820px;
+  height: 480px;
+  margin: auto 0;
+  -webkit-transform-style: preserve-3d;
+  transform-style: preserve-3d;
+  -webkit-perspective: 2400px;
+  perspective: 2400px;
+  cursor: pointer;
+}
+
+.card {
+  width: 100%;
+  height: 100%;
+  -webkit-transition: all 600ms;
+  transition: all 600ms;
+  -webkit-transform-style: preserve-3d;
+  transform-style: preserve-3d;
+}
+
+.card div {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  border: 3px solid #000;
+  font-weight: 400;
+  font-size: 64px;
+  text-align: center;
+  background: #FDFDFD;
+  -webkit-backface-visibility: hidden;
+  backface-visibility: hidden;
+  display: flex;
+  justify-content: center; 
+  align-items: center;
+  padding: 20px;
+}
+
+.card .back {
+  color: #000;
+  font-size: 32px;
+  -webkit-transform: rotateX(180deg);
+  transform: rotateX(180deg);
+}
+
+label:hover .card {
+  box-shadow: 0 20px 20px rgba(50, 50, 50, .2);
+  -webkit-transform: rotateX(20deg);
+  transform: rotateX(20deg);
+}
+
+input {
+  display: none;
+}
+
+:checked+.card {
+  transform: rotateX(180deg);
+  -webkit-transform: rotateX(180deg);
+}
+
+label:hover :checked+.card {
+  box-shadow: 0 20px 20px rgba(255, 255, 255, .2);
+  transform: rotateX(160deg);
+  -webkit-transform: rotateX(160deg);
+}
+
+.buttons__block {
+  display: flex;
+  justify-content: space-between;
+  gap: 20px;
+  margin-top: 40px;
+}
+
+.repeat {
+  width: 400px;
+}
+
+.repeat svg {
+  width: 34px;
+  height: 34px;
+  stroke: var(--danger-color);
+  transition: stroke 0.2s ease-out;
+}
+
+.repeat:active svg {
+  stroke: #FFFFFF;
+}
+
+.remember {
+  width: 400px;
+}
+
+.remember svg {
+  stroke: var(--success-color);
+  transition: stroke 0.2s ease-out;
+}
+
+.remember:active svg {
+  stroke: #FFFFFF;
+}
+
+.end__card {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+}
+
+.end__block {
+  position: relative;
+  width: 606px;
+  height: 435px;
+}
+
+.end__img {
+  width: 100%;
+  height: 100%;
+  background: url('/public/laptop.svg') no-repeat center / contain;
+}
+
+.end__text {
+  position: absolute;
+  top: 35%;
+  left: 35%;
+  margin: 0;
+  color: #000;
+  font-weight: 400;
+  font-size: 32px;
+  letter-spacing: -0.05em;
+  text-align: center;
+  transform: translate(-27%, -50%);
+}
+
+.end__button {
+  width: 610px;
+  margin-top: 40px;
+}
