+++ frontend/src/components/Loader/Loader.tsx
@@ -1,0 +1,15 @@
+import styles from './Loader.module.css';
+
+export const LoadingElement = () => {
+    return (
+        <div className={styles.element}>
+            <div className={styles.loading}>
+              <div></div>
+              <div></div>
+              <div></div>
+              <div></div>
+              <div></div>
+            </div>
+        </div>
+    );
+};
