+++ frontend/src/components/index.ts
@@ -1,3 +1,5 @@
 export { HomeLayout } from './layouts';
 export { Navbar } from './Navbar/Navbar';
 export { Folder } from './Folder/Folder';
+export { ExampleComponent } from './Cards/card'
+export { LoadingElement } from './Loader/Loader'
