+++ frontend/src/pages/CardPage/CardPage.module.css
@@ -18,184 +18,3 @@
 .back_button:hover .image {
   background: url('/public/back-hover.svg') no-repeat center / contain;
 }
-
-.card__block {
-  width: 820px;
-  margin: 0 auto;
-}
-
-.card {
-  position: relative;
-  display: block;
-  height: 480px;
-  cursor: pointer;
-}
-
-.front,
-.back {
-  position: absolute;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  border: 3px solid #000;
-  padding: 20px;
-  overflow: hidden;
-  background-color: #fdfdfd;
-  transition: transform .8s linear;
-  backface-visibility: hidden;
-}
-
-.front {
-  transform: perspective(1200px) rotateY(0deg);
-}
-
-.back {
-  transform: perspective(1200px) rotateY(180deg);
-}
-
-.card:hover>.front {
-  transform: perspective(1200px) rotateY(-180deg);
-}
-
-.card:hover>.back {
-  transform: perspective(1200px) rotateY(0deg);
-}
-
-.front__title {
-  margin: 0;
-  color: #000;
-  font-weight: 400;
-  font-size: 64px;
-  text-align: center;
-}
-
-.back__description {
-  margin: 0;
-  color: #000;
-  font-weight: 400;
-  font-size: 32px;
-  text-align: center;
-}
-
-.buttons__block {
-  display: flex;
-  justify-content: space-between;
-  gap: 20px;
-  margin-top: 40px;
-}
-
-.repeat {
-  width: 400px;
-  border: 3px solid var(--danger-color);
-  padding: 20px;
-  text-align: center;
-  background-color: #FDFDFD;
-  transition: background-color 0.2s ease-out;
-  cursor: pointer;
-}
-
-.repeat svg {
-  width: 34px;
-  height: 34px;
-  stroke: var(--danger-color);
-  transition: stroke 0.2s ease-out;
-}
-
-.repeat:hover {
-  background-color: #FFDDE4;
-}
-
-.repeat:active {
-  background-color: var(--danger-color);
-}
-
-.repeat:active svg {
-  stroke: #FFFFFF;
-}
-
-.remember {
-  width: 400px;
-  border: 3px solid var(--success-color);
-  padding: 20px;
-  text-align: center;
-  background-color: #FDFDFD;
-  transition: background-color 0.2s ease-out;
-  cursor: pointer;
-}
-
-.remember svg {
-  stroke: var(--success-color);
-  transition: stroke 0.2s ease-out;
-}
-
-.remember:hover {
-  background-color: #BAEEA7;
-}
-
-.remember:active {
-  background-color: var(--success-color);
-}
-
-.remember:active svg {
-  stroke: #FFFFFF;
-}
-
-.end__card {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-}
-
-.end__block {
-  position: relative;
-  width: 606px;
-  height: 435px;
-}
-
-.end__img {
-  width: 100%;
-  height: 100%;
-  background: url('/public/laptop.svg') no-repeat center / contain;
-}
-
-.end__text {
-  position: absolute;
-  top: 35%;
-  left: 35%;
-  margin: 0;
-  color: #000;
-  font-weight: 400;
-  font-size: 32px;
-  letter-spacing: -0.05em;
-  text-align: center;
-  transform: translate(-27%, -50%);
-}
-
-.end__button {
-  display: block;
-  width: 610px;
-  margin-top: 40px;
-  border: 5px solid var(--primary-color);
-  padding: 20px;
-  color: var(--primary-color);
-  font-weight: 500;
-  font-size: 32px;
-  letter-spacing: -0.05em;
-  text-align: center;
-  text-align: center;
-  background-color: #FDFDFD;
-  transition: background-color 0.2s ease-out, color 0.2s ease-out;
-  cursor: pointer;
-}
-
-.end__button:hover {
-  background-color: #D1D0E2;
-}
-
-.end__button:active {
-  background-color: var(--primary-color);
-  color: #ffffff;
-}
